Create pubmed.R
This file passes a query to the National Library of Medicine's PubMed site, and downloads 2 files. The first is an xml file that I hope to be able to pass through and build the references for. The second is a text file that contains the abstracts for all of the matching documents. 
Currently, the code stops and gives a warning if the count is too high (100).