Create BContin.R
This is an example of how I think about Bayes as it relates to biomedical tests. This function starts from the number of true and false positives and negatives, and calculates the PPV, NPV, Sn, SP, and Diagnostic Odds Ratio.

There could certainly be an adaptation of this function to take Sn, Sp, and Prevalence, and calculate everything else, I just haven't written it yet.